fix(keyhandler): shift modifier behavior (#1238)
Some keymaps in firenvim does not work as expected:

| Key pressed               | Firenvim keyhandler passed | Expected             | Notes                                      |
| ------------------------- | -------------------------- | -------------------- | ------------------------------------------ |
| `>` or `shift+,` in #1238 | `<S-lt>`                   | `<` or `<lt>`        | `<lt>` is used for `nvim_input`            |
| `\|` or `shift+\`         | `<S-Bar>`                  | `\|` or `<Bar>`      | should be parsed as `<Bar>` Same as above  |
| `ctrl+shift+/`            | `<C-S-?>`                  | `<C-S-/>` or `<C-?>` | depends on how firenvim calls `nvim_input` |
| `ctrl+shift+0`            | key press not detected     | `<C-S-0>` or `<C-)>` | depends on how firenvim calls `nvim_input` |
| `alt+shift+9`             | `<M-S-(>`                  | `<M-(>` or `<M-S-9>` | `<M-(>` is prefered in terminal nvim       |

Notes:

`evt.key` in key handler is the actual key the browser receives, and `evt.code` is the physical key on the keyboard.

For example, when pressing `shift+comma`, `evt.key` is `<`, and `evt.code` is `Comma`.

If we use `evt.code` to handle shift modifier, it would only work for some keyboard layout with same key-to-shiftedKey pairs (`:` to `;`, `/` to `?`, etc).

For some virtual keyboard layout (e.g., [programmers dvorak](https://github.com/ThePrimeagen/keyboards)), the above implementation may fail.

To fix this, we need a `keyToShiftedKey` mapping for mappable keys:

```
1!2@3#4$5%6^7&8*9(0)-_=+[{]}\|;:'",<.>/?`~
```

Then, we can maintain a reversed lookup dictionary (`shiftKeyToKey`) to find the unshifted key when handling shift modifier.

For example, `ctrl+shift+/` results in `evt.key=? event.ctrlKey=true event.shiftKey=true`. We can find `['?']: '/'` in `shiftKeyToKey` dictionary and then apply `nvim_input("<CS-/>")` or `nvim_input("<C-?>")`.

It is also noteworthy that `<C-S-/>` is a different keymap from `<C-?>` because `shift+/` does not necessarily trigger key `?` in all keyboard layouts.

This pull request does the following:

1. fixes #1238, and other `ctrl+shift`, `alt+shift`, `ctrl+alt+shift` combinations.
2. adds global options `keyboard_layouts` for specify key-shiftedKey pairs, `keyboard_layouts.default` is used as default layout.
3. adds global options `shiftAlt`, `shiftCtrl`, `shiftAltCtrl` to support custom shift behavior. For example, if `shiftAlt=true`, `ctrl+shift+/` will call `nvim_input("<CS-/>")`, else it will call `nvim_input("<C-?>")`.
4. WIP: add tests
4. TODO: support switching keyboard layouts